+++ boardgame-frontend/src/app/mypage/inquiries/list/page.tsx
@@ -1,58 +1,8 @@
-import Inquiry from "@/components/inquiry/Inquiry";
-import { useSession } from "next-auth/react";
+import InquiryList from "@/components/inquiry/InquiryList";
 import Image from "next/image";
 import Link from "next/link";
-import { useEffect, useState } from "react";
-
-interface Inquiry {
-  id: number;
-  title: string;
-  content: string;
-  answer: string | null;
-  createdAt: string;
-}
 
 export default function InquiriesListPage() {
-  const [inquiries, setInquiries] = useState<Inquiry[] | undefined>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const { data: session, status } = useSession();
-
-  // 내 문의 목록 API
-  const fetchInquiries = async () => {
-    if (!session?.user?.accessToken) {
-      return;
-    }
-    try {
-      setIsLoading(true);
-
-      const response = await fetch(`${process.env.NEXT_PUBLIC_API_SERVER_HOST}/api/inquiries/my`, {
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${session.user.accessToken}`,
-        },
-        credentials: "include",
-      });
-      if (!response.ok) {
-        throw new Error(`요청 실패: ${response.status}`);
-      }
-
-      const data = await response.json();
-      return data as Inquiry[];
-    } catch (error) {
-      console.error("데이터를 불러오지 못했습니다:", error);
-      throw error;
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchInquiries()
-      .then((data) => setInquiries(data))
-      .catch((err) => console.error(err));
-    console.log(inquiries);
-  }, []);
-
   return (
     <div className="pt-11 bg-white h-screen flex flex-col">
       <header className="flex px-5 py-[11px] gap-[2px] border-b border-[#F1F1F4]">
@@ -61,17 +11,9 @@ export default function InquiriesListPage() {
         </Link>
         <h1 className="font-semibold text-lg">1:1문의</h1>
       </header>
-      <main className="flex-1 overflow-y-auto">
-        {/* 문의 내역이 없을 때 보여주는 화면 */}
-        {inquiries?.length === undefined && (
-          <section className="flex flex-col items-center justify-center h-[calc(100%-100px)]">
-            <Image src="/icons/ic_empty_logo_with_bubble.svg" width={45} height={58} alt=""></Image>
-            <p className="mt-6 font-medium text-base text-center text-[#363636]">문의 내역이 없습니다.</p>
-          </section>
-        )}
-        {/* 문의 내역 */}
-        {inquiries && inquiries.map((inquiry: Inquiry) => <Inquiry key={inquiry.id} inquiry={inquiry} />)}
-      </main>
+
+      <InquiryList />
+
       <div className="fixed bottom-0 left-0 right-0 px-5 pb-10 pt-[18px] bg-white flex justify-center">
         <Link
           href="/mypage/inquiries/new"
